phase2-web-evolution/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
